--- pull_request_template.md
@@ -1,8 +1,0 @@
-### Is this a fix or a feature?
--
-
-### What is the change or fix?
--
-
-### How should this be tested?
--
