+++ src/assets/datasets/artistImages.js
