+++ src/assets/datasets/backupImages.js
@@ -1,0 +1,19 @@
+import cowbell1 from '../images/more-cowbells/cowbell1.jpeg';
+import cowbell2 from '../images/more-cowbells/cowbell2.jpeg';
+import cowbell3 from '../images/more-cowbells/cowbell3.jpeg';
+import cowbell4 from '../images/more-cowbells/cowbell4.jpeg';
+import cowbell5 from '../images/more-cowbells/cowbell5.jpeg';
+import cowbell6 from '../images/more-cowbells/cowbell6.jpeg';
+import cowbell7 from '../images/more-cowbells/cowbell7.jpeg';
+
+const backupImages = [
+  cowbell1,
+  cowbell2,
+  cowbell3,
+  cowbell4,
+  cowbell5,
+  cowbell6,
+  cowbell7
+];
+
+export default backupImages;
