--- src/documents/DTR
@@ -1,65 +1,0 @@
-Global State Management: My World Music
-Name subject to change
-
-Group Member Names: Bryan, Angie, Pete
-Goals and Expectations for the Project (What does each group member hope to get out of this project? What do we want to achieve as a team? How will we know that we're successful?):
-
-A good looking highly functional project that we're all proud to include in our professional portfolios.
-Get familiar with API key implementation
-Learn Redux
-How React and Redux work together
-Team strengths & collaboration styles (consider discussing your Pairin qualities here):
-
-Adaptability/Flexibility; (Angie and Pete)
-Determination/Positivity/Helpful; (Bryan)
-Available mentors for the struggle;
-Communication;
-An eye for design;
-How we can use our strengths to overcome obstacles:
-
-We will need to find balance between our communication to find the balance;
-Leverage our communication skills to make sure we remain in integrity
-Set daily goals => small accomplishable tasks
-Reach out to mentors
-Schedule Expectations (When are we available to work together and individually? What constraints do we have? Make sure to discuss the remote questions above in regards to this section):
-
-Angie: Must be done by 10pm; No regularly scheduled calls; Totally free to work any time otherwise;
-Pete: Available all the time;
-Bryan: More available in the morning and afternoon;
-All agreed hard cut off zoom by 9pm;
-Try to be exiting the project by 8pm
-Communication Expectations (How and how often will we communicate? How do we keep lines of communication open? How will we make decisions as a team? How will we communicate about our emotional and mental well-being with each other?):
-
-Lines of communication are open;
-Honestly, kindly, and directly;
-Work Flow = upgrading, refactoring, improving the code fluidly/openly
-Don't be afraid to revisit the dtr
-Abilities & Growth Expectations (Technical strengths and areas for desired improvement):
-
-Redux && React && Api keys
-UI/UX
-Accessibility
-Expanded Testing Vocabulary
-Growth: Form Elements/searching
-Loading/Performance of App/fetch rendering
-
-Workload Expectations (What features do we each want to work on?):
-We all want to pull our own weight/help each other
-Occasionally driver/navigator
-Workflow Expectations (Git workflow/Tools/Code Review/Reviewing Pull Requests/Debugging and Problem-solving Techniques):
-
-Decide tomorrow between merging vs. rebasing; Code reviews from each other, mentors, & rocks; PR template; Debugging/Problem solving - 30min max, then reach out to mentor/rock/instructor for help
-Pomodoro Break Schedule:
-
-45min on/5-7min off; After 2-3 hours, take longer break
-Expectations for giving and receiving feedback:
-
-Direct, kind, honest, quick
-This can be business professional
-Additional Remote Considerations:
-
-Zoom freezing - Try using Tuple/Slack/whatever we can; Disaster mode - Just call each other
-Project management tools we will use (GitHub projects or Trello are popular tools):
-
-GitHub Project board
-Leverage the project board to help stay on task
